+++ database/migrations/2018_01_29_145816_create_event_groups_table.php
@@ -13,18 +13,20 @@ class CreateEventGroupsTable extends Migration
      */
     public function up()
     {
-        Schema::create('event_groups', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('event_id');
-            $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
-            $table->string('title');
-            $table->timestamps();
+        if (!Schema::hasTable('event_groups')) {
+            Schema::create('event_groups', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('event_id');
+                $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
+                $table->string('title');
+                $table->timestamps();
 
-            $table->unique(['event_id', 'title'], 'event_group_title');
-        });
+                $table->unique(['event_id', 'title'], 'event_group_title');
+            });
 
-        DB::statement("ALTER TABLE event_groups ADD uuid BINARY(16) NULL AFTER id");
-        DB::statement('CREATE UNIQUE INDEX uuid_unique ON event_groups (uuid);');
+            DB::statement("ALTER TABLE event_groups ADD uuid BINARY(16) NULL AFTER id");
+            DB::statement('CREATE UNIQUE INDEX uuid_unique ON event_groups (uuid);');
+        }
     }
 
     /**
