+++ database/migrations/2018_02_14_144436_create_event_transcriptions_table.php
@@ -13,19 +13,21 @@ class CreateEventTranscriptionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('event_transcriptions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('classification_id');
-            $table->unique('classification_id');
-            $table->unsignedInteger('event_id');
-            $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
-            $table->unsignedInteger('group_id');
-            $table->foreign('group_id')->references('id')->on('event_groups')->onDelete('cascade');
-            $table->unsignedInteger('user_id');
-            $table->foreign('user_id')->references('id')->on('event_users')->onDelete('cascade');
+        if (! Schema::hasTable('event_transcriptions')) {
+            Schema::create('event_transcriptions', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('classification_id');
+                $table->unique('classification_id');
+                $table->unsignedInteger('event_id');
+                $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
+                $table->unsignedInteger('group_id');
+                $table->foreign('group_id')->references('id')->on('event_groups')->onDelete('cascade');
+                $table->unsignedInteger('user_id');
+                $table->foreign('user_id')->references('id')->on('event_users')->onDelete('cascade');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
